# lib/client/api/openapi.yaml
openapi: 3.0.0
info:
  description: |
    Nucmf_UECapabilityManagement Service.
    © 2020, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  title: Nucmf_UECapabilityManagement
  version: 1.0.2
...
servers:
- url: '{apiRoot}/nucmf-uecm/v1'
  variables:
    apiRoot:
      default: https://example.com
      description: apiRoot as defined in clause 4.4 of 3GPP TS 29.501
security:
- {}
- oAuth2ClientCredentials:
  - nucmf-uecm
paths:
...
  /dic-entries/{dicEntryId}:
    get:
      operationId: GetDicEntry
      parameters:
      - description: the ID of a dictionary entry in the UCMF
        explode: false
        in: path
        name: dicEntryId
        required: true
        schema:
          $ref: '#/components/schemas/DicEntryId'
        style: simple
      - description: Encoding format of of RAC Info
        explode: true
        in: query
        name: rac-format
        required: false
        schema:
          $ref: '#/components/schemas/RacFormat'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DicEntryData'
          description: Expected response to a valid request
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "406":
          description: 406 Not Acceptable
        "411":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Length Required
        "413":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Payload Too Large
        "415":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unsupported Media Type
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "501":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Implemented
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
      summary: Get an individual dictionary entry via dicEntryId
      tags:
      - dic Entry Id (Document)
...
components:
...
  schemas:
    DicEntryData:
      example:
        ueRadioCapability5GS:
          contentId: contentId
        ueRadioCapabilityEPS:
          contentId: contentId
        typeAllocationCode: typeAllocationCode
        manAssiUeRadioCapId: manAssiUeRadioCapId
        dicEntryId: 343953088
        plmnAssiUeRadioCapId: plmnAssiUeRadioCapId
      properties:
        dicEntryId:
          maximum: 4294967295
          minimum: 0
          type: integer
        typeAllocationCode:
          pattern: ^[0-9]{8}$
          type: string
        plmnAssiUeRadioCapId:
          format: byte
          type: string
        manAssiUeRadioCapId:
          format: byte
          type: string
        ueRadioCapability5GS:
          $ref: '#/components/schemas/RefToBinaryData'
        ueRadioCapabilityEPS:
          $ref: '#/components/schemas/RefToBinaryData'
      required:
      - typeAllocationCode
      type: object
...
    DicEntryId:
      maximum: 4294967295
      minimum: 0
      type: integer
...
    RacFormat:
      anyOf:
      - $ref: '#/components/schemas/RacFormat_anyOf'
      - type: string
    ProblemDetails:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        cause:
          type: string
        invalidParams:
          items:
            $ref: '#/components/schemas/InvalidParam'
          minItems: 1
          type: array
        supportedFeatures:
          pattern: ^[A-Fa-f0-9]*$
          type: string
      type: object
...
    InvalidParam:
      properties:
        param:
          type: string
        reason:
          type: string
      required:
      - param
      type: object
...
    RefToBinaryData:
      example:
        contentId: contentId
      properties:
        contentId:
          type: string
      required:
      - contentId
      type: object
...
    RacFormat_anyOf:
      enum:
      - 5GS
      - EPS
      type: string
  securitySchemes:
    oAuth2ClientCredentials:
      flows:
        clientCredentials:
          scopes:
            nucmf-uecm: Access to the Nucmf_UECapabilityManagement API
          tokenUrl: '{nrfApiRoot}/oauth2/token'
      type: oauth2