# lib/server/go/model_problem_details.go
/*
 * Nucmf_UECapabilityManagement
 *
 * Nucmf_UECapabilityManagement Service. © 2020, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved. 
 *
 * API version: 1.0.2
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ProblemDetails struct {

	Type string `json:"type,omitempty" bson:"type"`

	Title string `json:"title,omitempty" bson:"title"`

	Status int32 `json:"status,omitempty" bson:"status"`

	Detail string `json:"detail,omitempty" bson:"detail"`

	Instance string `json:"instance,omitempty" bson:"instance"`

	Cause string `json:"cause,omitempty" bson:"cause"`

	InvalidParams []InvalidParam `json:"invalidParams,omitempty" bson:"invalidParams"`

	SupportedFeatures string `json:"supportedFeatures,omitempty" bson:"supportedFeatures"`
}